Final panel [skip ci]
