+++ notebooks/MHWGPanels/design/final-panel.txt
@@ -1,0 +1,139 @@
+chr1 mh01USC-1pA 49 3.828
+chr1 mh01WL-110 60 4.324
+chr1 mh01WL-116 163 4.736
+chr1 mh01WL-077.v1 231 5.292
+chr1 mh01WL-083 209 4.127
+chr1 mh01WL-007 158 7.288
+chr1 mh01WL-010.v2 237 7.69
+chr1 mh01WL-087 71 4.719
+chr1 mh01WL-005.v1 169 11.555
+chr1 mh01WL-008.v2 258 6.129
+chr2 mh02WL-083 195 4.969
+chr2 mh02WL-088.v2 70 3.606
+chr2 mh02WL-067 242 5.16
+chr2 mh02LW-7 47 6.361
+chr2 mh02SCUZJ-0332125 259 5.015
+chr2 mh02WL-011 180 4.63
+chr2 mh02SHY-003.v1 195 4.166
+chr2 mh02KK-134.v3 104 5.206
+chr2 mh02WL-098 192 3.668
+chr2 mh02WL-002.v2 201 7.998
+chr2 mh02WL-003.v1 200 5.966
+chr3 mh03WL-006.v1 185 6.687
+chr3 mh03WL-046 52 3.855
+chr3 mh03WL-010 42 3.534
+chr3 mh03SHY-004 191 5.097
+chr3 mh03WL-012 171 3.763
+chr3 mh03WL-021.v1 168 4.953
+chr3 mh03WL-017.v2 231 5.865
+chr3 mh03WL-060 258 4.403
+chr3 mh03WL-058 208 5.085
+chr3 mh03USC-3qC.v2 257 10.74
+chr4 mh04WL-052.v1 243 13.685
+chr4 mh04SCUZJ-0098303 247 5.241
+chr4 mh04WL-031.v3 252 4.585
+chr4 mh04WL-072 233 3.627
+chr4 mh04WL-012.v1 79 4.364
+chr4 mh04LW-15 36 4.38
+chr4 mh04FHL-005.v1 133 5.697
+chr4 mh04SHY-001.v3 73 4.063
+chr4 mh04WL-028.v1 112 4.916
+chr4 mh04WL-044 82 4.288
+chr5 mh05KK-170.v3 256 11.777
+chr5 mh05WL-006.v1 175 3.843
+chr5 mh05WL-067.v1 171 4.563
+chr5 mh05USC-5qA 94 3.638
+chr5 mh05KK-178.v2 231 5.106
+chr5 mh05WL-049 251 4.379
+chr5 mh05WL-003.v1 89 3.515
+chr5 mh05WL-051 216 4.378
+chr5 mh05WL-047 239 4.025
+chr5 mh05WL-058.v1 190 5.02
+chr6 mh06WL-051 240 11.648
+chr6 mh06WL-005.v1 159 5.368
+chr6 mh06WL-068a 177 4.838
+chr6 mh06SCUZJ-0528857 167 10.552
+chr6 mh06WL-071.v1 187 3.901
+chr6 mh06WL-008.v1 156 5.759
+chr6 mh06WL-056 238 4.838
+chr6 mh06WL-047 96 3.354
+chr6 mh06USC-6qD 96 3.279
+chr7 mh07WL-088 127 3.951
+chr7 mh07KK-009.v3 171 9.324
+chr7 mh07WL-054 244 5.315
+chr7 mh07SCUZJ-0197487 242 5.547
+chr7 mh07WL-020.v1 199 4.244
+chr7 mh07SCUZJ-0380111.v2 110 4.725
+chr7 mh07WL-022.v2 238 8.556
+chr7 mh07SCUZJ-0502291.v1 225 7.47
+chr8 mh08WL-049 150 5.363
+chr8 mh08SHY-002 141 4.048
+chr8 mh08WL-040 240 5.003
+chr8 mh08USC-8pB 50 3.386
+chr8 mh08SHY-004 190 4.069
+chr8 mh08WL-068 84 3.388
+chr8 mh08WL-031 90 4.412
+chr8 mh08SCUZJ-0523897 80 3.851
+chr9 mh09WL-034 182 5.95
+chr9 mh09WL-044.v2 214 4.992
+chr9 mh09SHY-001.v1 173 6.227
+chr9 mh09WL-029 208 3.848
+chr9 mh09WL-020 100 6.871
+chr9 mh09WL-023.v2 64 4.554
+chr9 mh09USC-9qC 78 3.829
+chr10 mh10WL-031 238 9.692
+chr10 mh10SCUZJ-0082020 177 5.024
+chr10 mh10WL-052 170 4.281
+chr10 mh10WL-022.v1 144 3.569
+chr10 mh10WL-015 87 3.309
+chr10 mh10SCUZJ-0313839 212 4.637
+chr10 mh10SHY-002.v1 139 5.014
+chr10 mh10WL-028 80 3.432
+chr11 mh11WL-012 130 3.523
+chr11 mh11PK-63643.v1 37 4.016
+chr11 mh11WL-005.v1 190 4.403
+chr11 mh11WL-011.v1 193 3.612
+chr11 mh11WL-026 78 4.105
+chr11 mh11WL-023 47 4.644
+chr12 mh12KK-201.v2 177 9.58
+chr12 mh12USC-12qA 34 3.528
+chr12 mh12USC-12qB 78 4.04
+chr12 mh12WL-005.v2 184 7.294
+chr12 mh12WL-033 57 3.66
+chr12 mh12SCUZJ-0392651 201 6.723
+chr13 mh13USC-13qA.v2 142 8.644
+chr13 mh13KK-218.v5 253 8.862
+chr13 mh13KK-225.v3 207 3.714
+chr13 mh13KK-221.v1 253 8.327
+chr13 mh13USC-13qD 68 4.327
+chr14 mh14WL-036.v3 244 9.38
+chr14 mh14SCUZJ-0096560 196 5.323
+chr14 mh14SHY-003.v2 180 5.127
+chr14 mh14WL-028.v2 85 3.492
+chr15 mh15WL-004.v2 224 5.578
+chr15 mh15SCUZJ-0082880 192 10.059
+chr15 mh15WL-034 181 4.363
+chr15 mh15WL-001.v1 174 6.284
+chr15 mh15SCUZJ-0193223 144 4.025
+chr16 mh16WL-036 197 5.113
+chr16 mh16WL-003.v2 78 3.802
+chr16 mh16WL-009.v2 217 4.291
+chr16 mh16WL-012.v1 197 6.624
+chr16 mh16SHY-002 184 3.918
+chr17 mh17WL-019 90 3.662
+chr17 mh17WL-003 85 5.129
+chr17 mh17KK-272.v2 260 3.845
+chr17 mh17WL-037 102 3.531
+chr17 mh17FHL-005.v3 249 10.032
+chr18 mh18USC-18pA 77 3.424
+chr18 mh18WL-013 136 4.023
+chr18 mh18WL-018 51 3.594
+chr18 mh18WL-032 59 4.026
+chr19 mh19SHY-001.v4 67 3.731
+chr19 mh19USC-19qB.v2 106 5.113
+chr20 mh20WL-022.v2 252 6.37
+chr20 mh20WL-023.v3 93 3.933
+chr20 mh20WL-026.v2 97 3.752
+chr20 mh20USC-20qB 46 3.288
+chr22 mh22WL-006.v1 191 5.469
+chr22 mh22USC-22qB.v3 251 5.985
