+++ notebooks/MHWGPanels/design/prelim-panel.txt
@@ -1,0 +1,96 @@
+chr1 mh01WL-005.v1 169 11.555
+chr1 mh01WL-007 158 7.288
+chr1 mh01WL-008.v2 258 6.129
+chr1 mh01WL-010.v2 237 7.69
+chr1 mh01WL-077.v1 231 5.292
+chr1 mh01WL-087 71 4.719
+chr1 mh01WL-110 60 4.324
+chr2 mh02KK-134.v3 104 5.206
+chr2 mh02LW-7 47 6.361
+chr2 mh02SCUZJ-0332125 259 5.015
+chr2 mh02WL-002.v2 201 7.998
+chr2 mh02WL-003.v1 200 5.966
+chr2 mh02WL-067 242 5.16
+chr2 mh02WL-088.v2 70 3.606
+chr3 mh03SHY-004 191 5.097
+chr3 mh03USC-3qC.v2 257 10.74
+chr3 mh03WL-006.v1 185 6.687
+chr3 mh03WL-017.v2 231 5.865
+chr3 mh03WL-046 52 3.855
+chr3 mh03WL-058 208 5.085
+chr4 mh04FHL-005.v1 133 5.697
+chr4 mh04LW-15 36 4.38
+chr4 mh04SCUZJ-0098303 247 5.241
+chr4 mh04SHY-001.v3 73 4.063
+chr4 mh04WL-012.v1 79 4.364
+chr4 mh04WL-044 82 4.288
+chr4 mh04WL-052.v1 243 13.685
+chr5 mh05KK-170.v3 256 11.777
+chr5 mh05KK-178.v2 231 5.106
+chr5 mh05USC-5qA 94 3.638
+chr5 mh05WL-003.v1 89 3.515
+chr5 mh05WL-058.v1 190 5.02
+chr6 mh06SCUZJ-0528857 167 10.552
+chr6 mh06USC-6qD 96 3.279
+chr6 mh06WL-005.v1 159 5.368
+chr6 mh06WL-008.v1 156 5.759
+chr6 mh06WL-047 96 3.354
+chr6 mh06WL-051 240 11.648
+chr7 mh07KK-009.v3 171 9.324
+chr7 mh07SCUZJ-0197487 242 5.547
+chr7 mh07SCUZJ-0502291.v1 225 7.47
+chr7 mh07WL-022.v2 238 8.556
+chr7 mh07WL-054 244 5.315
+chr8 mh08SCUZJ-0523897 80 3.851
+chr8 mh08USC-8pB 50 3.386
+chr8 mh08WL-031 90 4.412
+chr8 mh08WL-040 240 5.003
+chr8 mh08WL-049 150 5.363
+chr8 mh08WL-068 84 3.388
+chr9 mh09SHY-001.v1 173 6.227
+chr9 mh09USC-9qC 78 3.829
+chr9 mh09WL-020 100 6.871
+chr9 mh09WL-023.v2 64 4.554
+chr9 mh09WL-034 182 5.95
+chr10 mh10SCUZJ-0082020 177 5.024
+chr10 mh10SHY-002.v1 139 5.014
+chr10 mh10WL-015 87 3.309
+chr10 mh10WL-028 80 3.432
+chr10 mh10WL-031 238 9.692
+chr11 mh11PK-63643.v1 37 4.016
+chr11 mh11WL-023 47 4.644
+chr11 mh11WL-026 78 4.105
+chr12 mh12KK-201.v2 177 9.58
+chr12 mh12SCUZJ-0392651 201 6.723
+chr12 mh12USC-12qA 34 3.528
+chr12 mh12USC-12qB 78 4.04
+chr12 mh12WL-005.v2 184 7.294
+chr12 mh12WL-033 57 3.66
+chr13 mh13KK-218.v5 253 8.862
+chr13 mh13KK-221.v1 253 8.327
+chr13 mh13USC-13qA.v2 142 8.644
+chr13 mh13USC-13qD 68 4.327
+chr14 mh14SCUZJ-0096560 196 5.323
+chr14 mh14SHY-003.v2 180 5.127
+chr14 mh14WL-028.v2 85 3.492
+chr14 mh14WL-036.v3 244 9.38
+chr15 mh15SCUZJ-0082880 192 10.059
+chr15 mh15WL-001.v1 174 6.284
+chr15 mh15WL-004.v2 224 5.578
+chr16 mh16WL-003.v2 78 3.802
+chr16 mh16WL-012.v1 197 6.624
+chr16 mh16WL-036 197 5.113
+chr17 mh17FHL-005.v3 249 10.032
+chr17 mh17WL-003 85 5.129
+chr17 mh17WL-019 90 3.662
+chr18 mh18USC-18pA 77 3.424
+chr18 mh18WL-018 51 3.594
+chr18 mh18WL-032 59 4.026
+chr19 mh19SHY-001.v4 67 3.731
+chr19 mh19USC-19qB.v2 106 5.113
+chr20 mh20USC-20qB 46 3.288
+chr20 mh20WL-022.v2 252 6.37
+chr20 mh20WL-023.v3 93 3.933
+chr20 mh20WL-026.v2 97 3.752
+chr22 mh22USC-22qB.v3 251 5.985
+chr22 mh22WL-006.v1 191 5.469
